desktop_app/requirements.txt
PyQt5
python-docx
docx2pdf
PyPDF2
cryptography
Pillow
Flask
reportlab
